tweet

+++ fake_twitter/include/fake_twitter/common.h
@@ -7,7 +7,7 @@ namespace fake_twitter {
 
     using PasswordHash = long;
     using Avatar = std::string;
-    using PKey = int;
+    using PKey = long;
     using Date = std::chrono::time_point<std::chrono::system_clock>;
     using HTML = std::string;
